Adapt per Nik's comments
